script: build: Resolve the issue of unable to compile using scripts
I found that it is being executed When ./rubikpi_builds.sh -a,
the script will be released quickly;I have identified the root cause of the issue.
In the setup_environment_select_RUBIKPi -> init_fuild_denv -> apply_poky_patches,
the wget command generates an error. In fact, this error can be ignored,
but due to the setting of set -e, the script cannot continue to execute downwards;
I don't want a similar situation to happen later, so I removed set -e

Signed-off-by: hongyang-rp <hongyang.zhao@thundersoft.com>

Change-Id: Ie7991bd8638818e6d00af2e650540af0b9c28ea1